+++ resources/cloudtrail_trails.go
@@ -2,7 +2,9 @@ package resources
 
 import (
 	"context"
+	"fmt"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
 	"github.com/aws/aws-sdk-go-v2/service/cloudtrail"
 	"github.com/aws/aws-sdk-go-v2/service/cloudtrail/types"
 	"github.com/cloudquery/cq-provider-aws/client"
@@ -27,10 +29,9 @@ func CloudtrailTrails() *schema.Table {
 				Resolver:    client.ResolveAWSAccount,
 			},
 			{
-				Name:        "region",
-				Description: "The AWS Region of the resource.",
-				Type:        schema.TypeString,
-				Resolver:    schema.PathResolver("HomeRegion"),
+				Name:        "tags",
+				Type:        schema.TypeJSON,
+				Description: "Any tags assigned to the resource",
 			},
 			{
 				Name:     "cloudwatch_logs_log_group_name",
@@ -39,7 +40,7 @@ func CloudtrailTrails() *schema.Table {
 			},
 			{
 				Name:        "is_logging",
-				Description: " Whether the CloudTrail is currently logging AWS API calls.",
+				Description: "Whether the CloudTrail is currently logging AWS API calls.",
 				Type:        schema.TypeBool,
 			},
 			{
@@ -113,13 +114,14 @@ func CloudtrailTrails() *schema.Table {
 				Type:        schema.TypeBool,
 			},
 			{
-				Name:        "home_region",
+				Name:        "region",
 				Description: "The region in which the trail was created.",
 				Type:        schema.TypeString,
+				Resolver:    schema.PathResolver("HomeRegion"),
 			},
 			{
 				Name:        "include_global_service_events",
-				Description: "Set to True to include AWS API calls from AWS global services such as IAM.",
+				Description: "Set to True to include AWS API calls from AWS global services such as IAM. Otherwise, False.",
 				Type:        schema.TypeBool,
 			},
 			{
@@ -134,7 +136,7 @@ func CloudtrailTrails() *schema.Table {
 			},
 			{
 				Name:        "kms_key_id",
-				Description: "Specifies the KMS key ID that encrypts the logs delivered by CloudTrail.",
+				Description: "Specifies the KMS key ID that encrypts the logs delivered by CloudTrail",
 				Type:        schema.TypeString,
 			},
 			{
@@ -144,33 +146,33 @@ func CloudtrailTrails() *schema.Table {
 			},
 			{
 				Name:        "name",
-				Description: "Name of the trail set by calling CreateTrail.",
+				Description: "Name of the trail set by calling CreateTrail",
 				Type:        schema.TypeString,
 			},
 			{
 				Name:        "s3_bucket_name",
-				Description: "Name of the Amazon S3 bucket into which CloudTrail delivers your trail files.",
+				Description: "Name of the Amazon S3 bucket into which CloudTrail delivers your trail files. See Amazon S3 Bucket Naming Requirements (https://docs.aws.amazon.com/awscloudtrail/latest/userguide/create_trail_naming_policy.html).",
 				Type:        schema.TypeString,
 			},
 			{
 				Name:        "s3_key_prefix",
-				Description: "Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery.",
+				Description: "Specifies the Amazon S3 key prefix that comes after the name of the bucket you have designated for log file delivery",
 				Type:        schema.TypeString,
 			},
 			{
 				Name:        "sns_topic_arn",
-				Description: "Specifies the ARN of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered.",
+				Description: "Specifies the ARN of the Amazon SNS topic that CloudTrail uses to send notifications when log files are delivered",
 				Type:        schema.TypeString,
 				Resolver:    schema.PathResolver("SnsTopicARN"),
 			},
 			{
 				Name:        "sns_topic_name",
-				Description: "This field is no longer in use.",
+				Description: "This field is no longer in use",
 				Type:        schema.TypeString,
 			},
 			{
 				Name:        "arn",
-				Description: "Specifies the ARN of the trail.",
+				Description: "Specifies the ARN of the trail",
 				Type:        schema.TypeString,
 				Resolver:    schema.PathResolver("TrailARN"),
 			},
@@ -178,7 +180,7 @@ func CloudtrailTrails() *schema.Table {
 		Relations: []*schema.Table{
 			{
 				Name:        "aws_cloudtrail_trail_event_selectors",
-				Description: "Use event selectors to further specify the management and data event settings for your trail.",
+				Description: "Use event selectors to further specify the management and data event settings for your trail",
 				Resolver:    fetchCloudtrailTrailEventSelectors,
 				Columns: []schema.Column{
 					{
@@ -189,23 +191,23 @@ func CloudtrailTrails() *schema.Table {
 					},
 					{
 						Name:        "trail_arn",
-						Description: "Specifies the ARN of the trail.",
+						Description: "Specifies the ARN of the trail",
 						Type:        schema.TypeString,
 						Resolver:    schema.ParentPathResolver("TrailARN"),
 					},
 					{
 						Name:        "exclude_management_event_sources",
-						Description: "An optional list of service event sources from which you do not want management events to be logged on your trail.",
+						Description: "An optional list of service event sources from which you do not want management events to be logged on your trail",
 						Type:        schema.TypeStringArray,
 					},
 					{
 						Name:        "include_management_events",
-						Description: "Specify if you want your event selector to include management events for your trail.",
+						Description: "Specify if you want your event selector to include management events for your trail",
 						Type:        schema.TypeBool,
 					},
 					{
 						Name:        "read_write_type",
-						Description: "Specify if you want your trail to log read-only events, write-only events, or all.",
+						Description: "Specify if you want your trail to log read-only events, write-only events, or all",
 						Type:        schema.TypeString,
 					},
 				},
@@ -226,13 +228,69 @@ func fetchCloudtrailTrails(ctx context.Context, meta schema.ClientMeta, parent *
 	if err != nil {
 		return err
 	}
-	res <- response.TrailList
+
+	processTrailsBundle := func(trails []types.Trail) error {
+		input := cloudtrail.ListTagsInput{
+			ResourceIdList: make([]string, 0, len(trails)),
+		}
+		for _, h := range trails {
+			input.ResourceIdList = append(input.ResourceIdList, *h.TrailARN)
+		}
+		for {
+			response, err := svc.ListTags(ctx, &input, func(options *cloudtrail.Options) {})
+			if err != nil {
+				return err
+			}
+			if len(response.ResourceTagList) == 0 {
+				break
+			}
+			for _, tr := range trails {
+				tags := getCloudTrailTagsByResourceID(*tr.TrailARN, response.ResourceTagList)
+				if len(tags) == 0 {
+					continue
+				}
+
+				wrapper := CloudTrailWrapper{
+					Trail: tr,
+					Tags:  make(map[string]interface{}, len(tags)),
+				}
+
+				for _, t := range tags {
+					wrapper.Tags[*t.Key] = t.Value
+				}
+				res <- wrapper
+			}
+
+			if aws.ToString(response.NextToken) == "" {
+				break
+			}
+			input.NextToken = response.NextToken
+		}
+		return nil
+	}
+
+	for i := 0; i < len(response.TrailList); i += 20 {
+		end := i + 10
+
+		if end > len(response.TrailList) {
+			end = len(response.TrailList)
+		}
+		trails := response.TrailList[i:end]
+		err := processTrailsBundle(trails)
+		if err != nil {
+			return err
+		}
+	}
+
 	return nil
 }
 func postCloudtrailTrailResolver(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource) error {
 	c := meta.(*client.Client)
 	svc := c.Services().Cloudtrail
-	r := resource.Item.(types.Trail)
+	r, ok := resource.Item.(CloudTrailWrapper)
+	if !ok {
+		return fmt.Errorf("expected CloudTrailWrapper but got %T", resource.Item)
+	}
 	response, err := svc.GetTrailStatus(ctx,
 		&cloudtrail.GetTrailStatusInput{Name: r.TrailARN}, func(o *cloudtrail.Options) {
 			o.Region = *r.HomeRegion
@@ -275,14 +333,18 @@ func postCloudtrailTrailResolver(ctx context.Context, meta schema.ClientMeta, re
 	}
 	return nil
 }
+
 func resolveCloudtrailTrailCloudwatchLogsLogGroupName(ctx context.Context, meta schema.ClientMeta, resource *schema.Resource, c schema.Column) error {
 	groupName := ""
 	log := meta.(*client.Client).Logger()
-	trail := resource.Item.(types.Trail)
-	if trail.CloudWatchLogsLogGroupArn != nil {
-		matches := client.GroupNameRegex.FindStringSubmatch(*trail.CloudWatchLogsLogGroupArn)
+	r, ok := resource.Item.(CloudTrailWrapper)
+	if !ok {
+		return fmt.Errorf("expected CloudTrailWrapper but got %T", resource.Item)
+	}
+	if r.CloudWatchLogsLogGroupArn != nil {
+		matches := client.GroupNameRegex.FindStringSubmatch(*r.CloudWatchLogsLogGroupArn)
 		if len(matches) < 2 {
-			log.Warn("CloudWatchLogsLogGroupARN doesn't fit standard regex", "arn", *trail.CloudWatchLogsLogGroupArn)
+			log.Warn("CloudWatchLogsLogGroupARN doesn't fit standard regex", "arn", *r.CloudWatchLogsLogGroupArn)
 		} else {
 			groupName = matches[1]
 		}
@@ -293,7 +355,10 @@ func resolveCloudtrailTrailCloudwatchLogsLogGroupName(ctx context.Context, meta 
 	return resource.Set("cloudwatch_logs_log_group_name", groupName)
 }
 func fetchCloudtrailTrailEventSelectors(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan interface{}) error {
-	r := parent.Item.(types.Trail)
+	r, ok := parent.Item.(CloudTrailWrapper)
+	if !ok {
+		return fmt.Errorf("expected CloudTrailWrapper but got %T", parent.Item)
+	}
 	c := meta.(*client.Client)
 	svc := c.Services().Cloudtrail
 	response, err := svc.GetEventSelectors(ctx, &cloudtrail.GetEventSelectorsInput{TrailName: r.TrailARN}, func(options *cloudtrail.Options) {
@@ -304,4 +369,18 @@ func fetchCloudtrailTrailEventSelectors(ctx context.Context, meta schema.ClientM
 	}
 	res <- response.EventSelectors
 	return nil
+}
+
+func getCloudTrailTagsByResourceID(id string, set []types.ResourceTag) []types.Tag {
+	for _, s := range set {
+		if *s.ResourceId == id {
+			return s.TagsList
+		}
+	}
+	return nil
+}
+
+type CloudTrailWrapper struct {
+	types.Trail
+	Tags map[string]interface{}
 }