+++ resources/elbv2_load_balancers_test.go
@@ -30,6 +30,12 @@ func buildElbv2LoadBalancers(t *testing.T, ctrl *gomock.Controller) client.Servi
 		gomock.Any(),
 	).Return(fakeLoadBalancerAttributes(), nil)
 
+	tags := elasticloadbalancingv2.DescribeTagsOutput{}
+	err = faker.FakeData(&tags)
+	if err != nil {
+		t.Fatal(err)
+	}
+	m.EXPECT().DescribeTags(gomock.Any(), gomock.Any(), gomock.Any()).Return(&tags, nil)
 	return client.Services{
 		ELBv2: m,
 	}